pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "self-healing-agent"
version = "0.1.0"
description = "Local observability agent — observe, detect, diagnose, act (ports & adapters)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "psutil>=5.9.0",
  "pyyaml>=6.0",
  "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.8.0",
  "pre-commit>=3.0.0",
]

[project.scripts]
sha-agent = "self_healing_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
